requirements.txt
stomp.py
pandas
requests
python-dotenv